+++ packages/waku/src/lib/plugins/vite-plugin-rsc-serve.ts
@@ -26,7 +26,7 @@ export function rscServePlugin(opts: {
   distDir: string;
   distPublic: string;
   srcServeFile: string;
-  serve: 'netlify' | 'cloudflare' | 'partykit' | 'deno' | 'aws-lambda';
+  serve: 'partykit' | 'deno' | 'aws-lambda';
 }): Plugin {
   return {
     name: 'rsc-serve-plugin',
